Добавил хук useTheme.
